Added process to handle statements (foreach / if for example)
The "Just Keep Swinging" variable counts up and down based on known patterns for starting if and foreach statements (if someone deviates it wont work, or we can expand as needed)

Part of this process also handles the other option of a statement being written all on one line.

NB: this will not work with multiline ForEach Method.